refactor (sample): extract all scribble demo entry to ScribbleDemoActivity
