+++ app/sample/src/main/res/layout/activity_main.xml
@@ -59,28 +59,10 @@
             android:text="Pen Stylus WebView Demo" />
 
         <Button
-            android:id="@+id/button_scribble_touch_helper"
+            android:id="@+id/button_scribble_demo"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Scribble Stylus Touch Helper Demo" />
-
-        <Button
-            android:id="@+id/button_surfaceview_stylus_scribble"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Scribble Stylus SurfaceView Demo"/>
-
-        <Button
-            android:id="@+id/button_webview_stylus_scribble"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Scribble Stylus WebView Demo"/>
-
-        <Button
-            android:id="@+id/button_touch_screen_scribble"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Scribble Touch Screen Demo"/>
+            android:text="Scribble Demo" />
 
         <Button
             android:id="@+id/btn_onyx_test"
