+++ app/sample/src/main/res/layout/activity_sribble_demo.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical" android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin">
+    <Button
+        android:id="@+id/button_scribble_touch_helper"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Scribble Stylus Touch Helper Demo" />
+
+    <Button
+        android:id="@+id/button_surfaceview_stylus_scribble"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Scribble Stylus SurfaceView Demo"/>
+
+    <Button
+        android:id="@+id/button_webview_stylus_scribble"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Scribble Stylus WebView Demo"/>
+
+    <Button
+        android:id="@+id/button_touch_screen_scribble"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Scribble Touch Screen Demo"/>
+
+</LinearLayout>
